add channel home and videos component, update navbar to update route in state
