+++ src/modules/channel/subcomponents/channel_home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+class ChannelHome extends React.Component {
+
+  render() {
+    return (
+      <div>Home</div>
+    );
+  }
+}
+
+export { ChannelHome };
