+++ src/modules/channel/subcomponents/index.js
@@ -1,2 +1,4 @@
 export * from './navbar';
 export * from './channel_videos';
+export * from './channel_home';
+export * from './featured_video';
